Use message count to avoid re-processing errors
If a retry fails, it will go back into the error queue, but we don't
want it to immediately be processed again. By using a `times` count
based on the number of messages in the queue, rather than a loop in
combination with the `pop` triplet check, we will avoid continuously
re-processing the errors.