Add Max Workers limit to emque-consuming config and control
Why? To summarize the reasons mentioned in the pull request, the current
setup for emque-consuming basically allows for worker inflation to the
point of breakage. This adds a sane, if arbitrary, limit to the number
of workers that can be working at once. It also allows for a config
setting that users can set to override the default.